+++ lib/src/store.dart
@@ -1,20 +1,29 @@
 import 'dart:async';
 
+import 'package:disposebag/disposebag.dart';
 import 'package:meta/meta.dart';
-import 'package:rx_redux/rx_redux.dart';
+
+import '../rx_redux.dart';
 
 /// Rx Redux Store.
 /// Redux store based on [Stream].
 class RxReduxStore<A, S> {
-  final _action = StreamController<A>.broadcast();
+  final void Function(A) _dispatch;
 
-  StreamSubscription _subscription;
-  S _state;
-  Stream<S> _stateStream;
-  bool Function(S previous, S next) _equals;
+  final GetState<S> _getState;
+  final Stream<S> Function() _stateStream;
+
+  final Future<void> Function() _dispose;
+
+  const RxReduxStore._(
+    this._dispatch,
+    this._getState,
+    this._stateStream,
+    this._dispose,
+  );
 
   /// Construct a [RxReduxStore]
-  RxReduxStore({
+  factory RxReduxStore({
     @required S initialState,
     @required List<SideEffect<A, S>> sideEffects,
     @required Reducer<A, S> reducer,
@@ -22,19 +31,35 @@ class RxReduxStore<A, S> {
     bool Function(S previous, S next) equals,
     RxReduxLogger logger,
   }) {
-    _state = initialState;
-    _equals = equals;
+    final actionSubject = StreamController<A>.broadcast(sync: true);
 
-    _stateStream = _action.stream.reduxStore<S>(
+    final stateStream = actionSubject.stream.reduxStore<S>(
       initialStateSupplier: () => initialState,
       sideEffects: sideEffects,
       reducer: reducer,
       logger: logger,
     );
 
-    _subscription = _stateStream.listen(
-      (state) => _state = state,
-      onError: handleError,
+    var currentState = initialState;
+
+    return RxReduxStore._(
+      actionSubject.add,
+      () => currentState,
+      () => () async* {
+        yield currentState;
+        yield* stateStream;
+      }()
+          .distinct(equals),
+      DisposeBag(
+        [
+          stateStream.listen(
+            (newState) => currentState = newState,
+            onError: handleError,
+          ),
+          actionSubject
+        ],
+        false,
+      ).dispose,
     );
   }
 
@@ -43,83 +68,14 @@ class RxReduxStore<A, S> {
   /// The stream skips states if they are equal to the previous state.
   /// Equality is determined by the provided [equals] method. If that is omitted,
   /// the '==' operator is used.
-  Stream<S> get stateStream {
-    final stream = () async* {
-      yield _state;
-      yield* _stateStream;
-    }();
-    return stream.distinct(_equals);
-  }
+  Stream<S> get stateStream => _stateStream();
 
   /// Get current state synchronously.
-  S get state => _state;
+  S get state => _getState();
 
   /// Dispatch action to store.
-  void dispatch(A action) => _action.add(action);
+  void dispatch(A action) => _dispatch(action);
 
   /// Dispose all resources.
-  Future<void> dispose() async {
-    await _subscription.cancel();
-    await _action.close();
-  }
-}
-
-enum _Action { a1, a2, a3, b1, b2, b3, b0 }
-
-void main() async {
-  final store = RxReduxStore<_Action, int>(
-    initialState: 0,
-    sideEffects: [
-      (action, getState) => action
-          .where((event) => event == _Action.a1)
-          .asyncExpand((event) =>
-              Stream.periodic(const Duration(seconds: 1), (_) => _Action.b1)
-                  .take(1)),
-      (action, getState) => action
-          .where((event) => event == _Action.a2)
-          .asyncExpand((event) => Stream.periodic(
-              const Duration(milliseconds: 500), (_) => _Action.b2).take(1)),
-      (action, getState) => action
-          .where((event) => event == _Action.a3)
-          .asyncExpand((event) => Stream.periodic(
-              const Duration(milliseconds: 300), (_) => _Action.b3).take(1)),
-      (action, getState) => Stream.error(Exception()),
-    ],
-    reducer: (state, action) {
-      switch (action) {
-        case _Action.b1:
-          return state + 1;
-        case _Action.b2:
-          return state + 2;
-        case _Action.b3:
-          return state + 3;
-        default:
-          return state;
-      }
-    },
-    handleError: (e, s) => print('Oh no $e'),
-  );
-
-  await Future.delayed(const Duration(seconds: 1));
-  print('<~> [1] ${store.state}');
-  final sub = store.stateStream.listen((event) => print('~> [1] $event'));
-
-  store.dispatch(_Action.a1);
-  store.dispatch(_Action.a2);
-  store.dispatch(_Action.a3);
-
-  await Future.delayed(const Duration(seconds: 1));
-
-  store.dispatch(_Action.a1);
-  store.dispatch(_Action.a2);
-  store.dispatch(_Action.a3);
-
-  await Future.delayed(const Duration(seconds: 1));
-  store.dispatch(_Action.b0);
-  store.dispatch(_Action.b0);
-  store.dispatch(_Action.b0);
-
-  print('dispose');
-  await sub.cancel();
-  await store.dispose();
+  Future<void> dispose() => _dispose();
 }
